requirements.txt
pillow
numpy
matplotlib
tensorflow
requests
lxml
PyQt5